refactor: board dto 패키지 생성 후 이동 및 Diary 좋아요 컬럼 추가 후 정렬
- Diary 좋아요 개수 컬럼 및 인덱스 추가
- Diary 쓰기에서 이전보다 시간이 걸리겠지만 읽기에서는 속도 개선됨

+++ src/main/java/com/phcworld/repository/board/dto/DiarySelectDto.java
@@ -25,7 +25,7 @@ public class DiarySelectDto {
 
     private Long countOfAnswers;
 
-    private Integer countOfGood;
+    private Long countOfGood;
 
     private LocalDateTime updateDate;
